app_production/requirements.txt
pyspark==3.5.1
kafka-python==2.0.2
pandas==2.0.3
boto3==1.28.57
requests==2.31.0
matplotlib==3.7.2
notebook==7.0.3
python-dotenv==1.0.0
findspark==2.0.1
selenium==4.15.0
beautifulsoup4==4.12.2
webdriver-manager==4.0.1

# Google Cloud
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
google-cloud-secret-manager==2.16.4

# OpenAI / LLM
openai>=1.12.0

# Spark / Delta Live Tables support
# delta-spark incompatível com pyspark 3.5 no momento
delta-spark==3.1.0
reportlab==3.6.13

# Databricks SQL connector (para consultas e geração de gráficos)
databricks-sql-connector==3.1.1

# Databricks SDK
databricks-sdk==0.18.0

# Terraform provider (para pipeline deployment)
python-terraform==0.10.1
